+++ Assets/Plugins/LeanAudio.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 52e41e970d9353942b27458440bec9eb
+timeCreated: 1427917971
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
