+++ Assets/Plugins/LeanTest.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 82464f26ca2ba284a8f92f51248c574a
+timeCreated: 1427917971
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
